+++ programs/dynamic-bonding-curve/src/event.rs
@@ -190,6 +190,13 @@ pub struct EvtPartnerWithdrawMigrationFee {
 #[event]
 pub struct EvtClaimPoolCreationFee {
     pub pool: Pubkey,
-    pub treasury: Pubkey,
+    pub receiver: Pubkey,
+    pub creation_fee: u64,
+}
+
+#[event]
+pub struct EvtPartnerClaimPoolCreationFee {
+    pub pool: Pubkey,
+    pub partner: Pubkey,
     pub creation_fee: u64,
 }