+++ programs/dynamic-bonding-curve/src/instructions/initialize_pool/ix_initialize_virtual_pool_with_spl_token.rs
@@ -15,6 +15,7 @@ use crate::{
     cpi_checker::cpi_with_account_lamport_and_owner_checking,
     process_create_token_metadata,
     state::{fee::VolatilityTracker, PoolConfig, PoolType, TokenType, VirtualPool},
+    token::transfer_lamports_from_user,
     EvtInitializePool, PoolError, ProcessCreateTokenMetadataParams,
 };
 
@@ -207,6 +208,16 @@ pub fn handle_initialize_virtual_pool_with_spl_token<'c: 'info, 'info>(
         token_mint_authority,
     )?;
 
+    // charge pool creation fee
+    if config.pool_creation_fee > 0 {
+        transfer_lamports_from_user(
+            ctx.accounts.payer.to_account_info(),
+            ctx.accounts.pool.to_account_info(),
+            ctx.accounts.system_program.to_account_info(),
+            config.pool_creation_fee,
+        )?;
+    }
+
     // init pool
     let mut pool = ctx.accounts.pool.load_init()?;
 
@@ -223,7 +234,6 @@ pub fn handle_initialize_virtual_pool_with_spl_token<'c: 'info, 'info>(
         PoolType::SplToken.into(),
         activation_point,
         initial_base_supply,
-        false,
     );
 
     emit_cpi!(EvtInitializePool {