+++ programs/dynamic-bonding-curve/src/instructions/initialize_pool/ix_initialize_virtual_pool_with_token2022.rs
@@ -1,7 +1,6 @@
 use super::InitializePoolParameters;
 use super::{max_key, min_key};
-use crate::constants::fee::TOKEN_2022_POOL_WITH_OUTPUT_FEE_COLLECTION_CREATION_FEE;
-use crate::state::CollectFeeMode;
+use crate::token::transfer_lamports_from_user;
 use crate::{
     activation_handler::get_current_point,
     const_pda,
@@ -12,8 +11,6 @@ use crate::{
     EvtInitializePool, PoolError,
 };
 use anchor_lang::prelude::*;
-use anchor_lang::solana_program::program::invoke;
-use anchor_lang::solana_program::system_instruction;
 use anchor_spl::token_2022::spl_token_2022::instruction::AuthorityType;
 use anchor_spl::token_interface::spl_pod::optional_keys::OptionalNonZeroPubkey;
 use anchor_spl::{
@@ -150,26 +147,6 @@ pub fn handle_initialize_virtual_pool_with_token2022<'c: 'info, 'info>(
     );
     token_metadata_initialize(cpi_ctx, name, symbol, uri)?;
 
-    let collect_fee_mode = CollectFeeMode::try_from(config.collect_fee_mode)
-        .map_err(|_| PoolError::InvalidCollectFeeMode)?;
-
-    let should_charge_creation_fee = collect_fee_mode == CollectFeeMode::OutputToken;
-
-    if should_charge_creation_fee {
-        invoke(
-            &system_instruction::transfer(
-                &ctx.accounts.payer.key(),
-                &ctx.accounts.pool.key(),
-                TOKEN_2022_POOL_WITH_OUTPUT_FEE_COLLECTION_CREATION_FEE,
-            ),
-            &[
-                ctx.accounts.payer.to_account_info(),
-                ctx.accounts.pool.to_account_info(),
-                ctx.accounts.system_program.to_account_info(),
-            ],
-        )?;
-    }
-
     // transfer minimum rent to mint account
     update_account_lamports_to_minimum_balance(
         ctx.accounts.base_mint.to_account_info(),
@@ -251,6 +228,16 @@ pub fn handle_initialize_virtual_pool_with_token2022<'c: 'info, 'info>(
         token_mint_authority,
     )?;
 
+    // charge pool creation fee
+    if config.pool_creation_fee > 0 {
+        transfer_lamports_from_user(
+            ctx.accounts.payer.to_account_info(),
+            ctx.accounts.pool.to_account_info(),
+            ctx.accounts.system_program.to_account_info(),
+            config.pool_creation_fee,
+        )?;
+    }
+
     // init pool
     let mut pool = ctx.accounts.pool.load_init()?;
 
@@ -267,7 +254,6 @@ pub fn handle_initialize_virtual_pool_with_token2022<'c: 'info, 'info>(
         PoolType::Token2022.into(),
         activation_point,
         initial_base_supply,
-        should_charge_creation_fee,
     );
 
     emit_cpi!(EvtInitializePool {