+++ programs/dynamic-bonding-curve/src/instructions/operator/mod.rs
@@ -1,0 +1,8 @@
+pub mod ix_claim_protocol_fee;
+pub use ix_claim_protocol_fee::*;
+pub mod ix_withdraw_protocol_surplus;
+pub use ix_withdraw_protocol_surplus::*;
+pub mod ix_claim_protocol_pool_creation_fee;
+pub use ix_claim_protocol_pool_creation_fee::*;
+pub mod ix_claim_legacy_pool_creation_fee;
+pub use ix_claim_legacy_pool_creation_fee::*;