+++ tests/fee_swap.tests.ts
@@ -107,6 +107,7 @@ describe("Fee Swap test", () => {
           dynamicFee: 0,
           poolFeeBps: 0,
         },
+        poolCreationFee: new BN(0),
         padding: [],
         curve: curves,
       };
@@ -433,6 +434,7 @@ describe("Fee Swap test", () => {
           dynamicFee: 0,
           poolFeeBps: 0,
         },
+        poolCreationFee: new BN(0),
         padding: [],
         curve: curves,
       };