+++ tests/full_flow_with_sol.tests.ts
@@ -133,6 +133,7 @@ describe("Full flow with spl-token", () => {
         dynamicFee: 0,
         poolFeeBps: 0,
       },
+      poolCreationFee: new BN(0),
       padding: [],
       curve: curves,
     };