+++ tests/migrate_to_damm_v2.tests.ts
@@ -124,6 +124,7 @@ describe("Migrate to damm v2", () => {
         dynamicFee: 0,
         poolFeeBps: 0,
       },
+      poolCreationFee: new BN(0),
       padding: [],
       curve: curves,
     };