+++ tests/simulate_cu_swap.tests.ts
@@ -91,6 +91,7 @@ describe("Simulate CU swap", () => {
           dynamicFee: 0,
           poolFeeBps: 0,
         },
+        poolCreationFee: new BN(0),
         padding: [],
         curve: curves,
       };