+++ tests/token_authority.tests.ts
@@ -506,6 +506,7 @@ async function createPool(
       dynamicFee: 0,
       poolFeeBps: 0,
     },
+    poolCreationFee: new BN(0),
     padding: [],
     curve: curves,
   };